# openapi/docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IHerb solution server",
        "description": "API for server of solution for hackaton",
        "version": "0.0.1",
        "contact": {}
    },
    "tags": [],
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
            "LoginDto": {
                "type": "object",
                "properties": {}
            }
        }
    },
    "paths": {
        "/user/login": {
            "post": {
                "operationId": "UserController_login",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/user/{vitamin}": {
            "put": {
                "operationId": "UserController_updateVitamin",
                "parameters": [
                    {
                        "name": "vitamin",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/user/recommendations": {
            "get": {
                "operationId": "UserController_getRecommendations",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/user/contraindications": {
            "get": {
                "operationId": "UserController_getContraindications",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/product/{id}": {
            "get": {
                "operationId": "ProductController_getProduct",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        }
    }
}